Revert "Merge pull request #7 from sitloboi2012/main"
This reverts commit 3b8873b925d88059071f3d0350f5b6298d3a8179, reversing
changes made to ab5963e7629cba8d2883fed6181e3fcb839e9c61.
