+++ .gitignore
@@ -1,3 +1,1 @@
-.streamlit/
-.vscode/
-*/__pycache__/
+./streamlit/secrets.toml
