+++ requirements.txt
@@ -2,4 +2,3 @@ openai==0.28.1
 langchain==0.0.325
 pypdf2==3.0.1
 PyMuPDF==1.23.5
-pdfplumber
